test: split the suppression-provenance tests into their own module
CodeFactor moved from "1 issue" to "2 issues" on the previous commit — the
first signal it has given that tracks a change of mine, since it publishes
a count with no reachable detail. Differencing pylint (all checkers, plus
the extensions CodeFactor enables) against the merge base identifies the
new one: `C0302: Too many lines in module (1005/1000)`, caused by the
justification test added there.

Fixed the way this repository asks for a size problem to be fixed — move
responsibility out to a properly-owned module, never trim to fit
(`AGENTS.md`). The suppression-provenance tests share one subject (ADR-067
D3's rule record and how each renderer states it) distinct from the
format-invariant subject the parent module owns, so they become
`tests/test_no_baseline_suppression_provenance.py`. The JUnit exit-axis and
published-schema tests stayed behind: those are about the parent's subject,
not this one. Test count goes 119 -> 122; nothing was dropped in the move.

Also fixes `C0415` (the schema-path import moved to module scope). The
remaining new-vs-base findings are all deliberate or mechanical: four
`C1805`s are the explicit `assert exit_code != 0` guards that exist because
an implicit-boolean `if exit_code:` made two assertions unreachable, and
the `C0411`/`C0413` pair is this repo's established `sys.path` +
`example_catalog` test idiom, already present in the parent module and
inherited by the split.

Whether `C0302` was CodeFactor's second issue is still a hypothesis — the
check gives no detail — but it is the only new finding whose appearance
coincides with the count moving, and crossing a module-size cap is worth
fixing on its own terms.
